+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased]
 
+### Added
+- Content Inset support
+
 ## [1.3.0](https://github.com/piemapping/pie-overlay-menu-ios/tree/1.3.0) (2017-02-14)
 ### Added
 - Support for status bar style for both content and menu